Moved constants from siirtc header to dedicated constants file
